+++ main.py
@@ -5,7 +5,7 @@
 
 您可以免费地使用、修改和分发PyInst ，但必须包含原始作者的版权声明和下列许可声明。
 
-You may use, modify, and distribute PyInst free of charge, but you must include
+You may use, modify, and distribute pyinsts free of charge, but you must include
 the original author's copyright notice and the following license notice.
 """
 
@@ -35,8 +35,8 @@ SOFTWARE.
 
 
 # 这是一个示例 Python 脚本。
-import common
-import libs
+from src.pyinsts import common
+from src.pyinsts import libs
 
 fswp_addr = 'TCPIP::192.168.1.1::INSTR'
 n9020b_addr = 'TCPIP::192.168.1.2::INSTR'